safari-600.7-branch/LayoutTests/fast/hidpi/image-srcset-only-src-attribute.html

<head>
<html>
<script src="resources/srcset-helper.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
    }, false);
</script>
</head>
<body id="body">
    <div>This test passes if this img tag below is a green square when the scale factor is 2. It ensures that the selection algorithm does not change the behavior of the src attribute, even with a scale factor greater than 1</div>
    <img id="foo" src="resources/green-200-px-square.png"></img>
</body>
</html>
