LayoutTests/fast/hidpi/image-srcset-only-src-attribute-expected.txt

PASS document.getElementById("foo").clientWidth==200 is true
This test passes if this img tag below is a green square when the scale factor is 2. It ensures that the selection algorithm does not change the behavior of the src attribute, even with a scale factor greater than 1
